refactor(Site): :recycle: Make reusable the function bindLinksWithSiteIds
