+++ src/pages/Sites/Sites.interfaces.ts
@@ -1,5 +1,5 @@
 import { SiteResponse } from '@API/REST.interfaces';
 
-export interface SiteExtended extends SiteResponse {
-  connected: string[];
+export interface SiteWithLinks extends SiteResponse {
+  targetIds: string[];
 }
